Refactor texture sampling and bump mapping
- Added support for normal scaling in the RayTracingMaterial struct
- Modified the sampleMap function in texture_sampling.fs to apply the normal scaling to the normal map
- Commented out the bump mapping code in perturbNormal function in texture_sampling.fs

This commit refactors the texture sampling and bump mapping functionality in the shaders. It introduces support for normal scaling in the RayTracingMaterial struct, allowing for more control over the appearance of normal maps. The sampleMap function in texture_sampling.fs has been modified to apply the normal scaling to the normal map. Additionally, the bump mapping code in the perturbNormal function has been commented out for now.